Started the basic microscope application framework
A system of callbacks is used to transition the microscope through
a few basic states: currently just Off and Holding.  There are
also callbacks to help mediate a "Panic" shutdown.

When the application starts, there is an initialization phase where
callbacks are registered and default configurations can be loaded.
The system then transitions through the Off state to the Holding state.

In the Off state, devices are configured/configurable but otherwise
shutdown (and potentially unallocated).

In the Holding state, devices have been initialized and are waiting
to get used during some microscope task.

TODO:
 - add device-galvo
 - ideas for configuration system
 - ideas for channel mapping
 - add a basic task